requirements.txt
Flask==3.1.2
opencv-python==4.12.0.88
numpy==2.2.6
Pillow>=10.0.0
scikit-learn>=1.6.1
joblib>=1.5.2

